AI_Shopping_Helper/requirements.txt
# AI Shopping Helper - Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (SQLite3 is built into Python)
# No additional database dependencies required

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
scrapy==2.11.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
pytesseract==0.3.10

# AI/ML Libraries
scikit-learn==1.3.1
sentence-transformers==2.7.0
chromadb==0.4.15

# HTTP Client
aiohttp==3.8.5

# Configuration
python-dotenv==1.0.0

# Development Tools
pytest==7.4.2
pytest-asyncio==0.21.1

# Logging
structlog==23.1.0

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Caching
redis==5.0.0

# Environment Variables
python-decouple==3.8
